+++ lib/rupert/rpm/header.rb
@@ -25,6 +25,7 @@ module Rupert
         :vendor             => 1011,
         :url                => 1020,
         :source_rpm         => 1044,
+        :group              => 1016,
       }.freeze
 
       # Creates a new header.